+++ thornode_bot.py
@@ -1,39 +1,19 @@
 import atexit
-import os
-import logging
-import requests
-import json
-import subprocess
-from datetime import datetime
 
 from telegram.ext.dispatcher import run_async
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.ext import (
     Updater,
     CommandHandler,
     PicklePersistence,
     CallbackQueryHandler,
     MessageHandler,
-    ConversationHandler,
     Filters
 )
 
-# Enable logging
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
-logger = logging.getLogger(__name__)
+from constants import *
+from helpers import *
+from jobs import *
 
-"""
-######################################################################################################################################################
-Static & environment variables
-######################################################################################################################################################
-"""
-
-TELEGRAM_BOT_TOKEN = os.environ['TELEGRAM_BOT_TOKEN']
-THORCHAIN_NODE_IP = os.environ['THORCHAIN_NODE_IP'] if 'THORCHAIN_NODE_IP' in os.environ else 'localhost'
-DEBUG = bool(os.environ['DEBUG'] == 'True') if 'DEBUG' in os.environ else False
-ADMIN_USER_IDS = [int(admin_id) for admin_id in
-                  os.environ['ADMIN_USER_IDS'].split(",")] if 'ADMIN_USER_IDS' in os.environ else []
-DOCKER_CURL_CMD = "curl --max-time 30 --no-buffer -s --unix-socket /var/run/docker.sock"
 
 """
 ######################################################################################################################################################
@@ -362,425 +342,9 @@ def keep_container_running(update, context):
 
 """
 ######################################################################################################################################################
-Jobs
-######################################################################################################################################################
-"""
-
-
-def thornode_checks(context):
-    """
-    Periodic checks of various node stats
-    """
-
-    check_thornodes(context)
-    check_thorchain_block_height(context)
-    check_thorchain_catch_up_status(context)
-    check_thorchain_midgard_api(context)
-
-
-def check_thornodes(context):
-    """
-    Check all added thornodes for any changes.
-    """
-
-    chat_id = context.job.context['chat_id']
-    user_data = context.job.context['user_data']
-
-    # Flag to show home buttons or not
-    message_sent = False
-
-    # List to delete entries after loop
-    delete_addresses = []
-
-    # Iterate through all keys
-    for address in user_data['nodes'].keys():
-        remote_node = get_thornode_object(address=address)
-        local_node = user_data['nodes'][address]
-
-        if remote_node is None:
-            text = 'THORNode is not active anymore! 💀' + '\n' + \
-                   'Address: ' + address + '\n\n' + \
-                   'Please enter another THORNode address.'
-
-            delete_addresses.append(address)
-
-            # Send message
-            context.bot.send_message(chat_id, text)
-            message_sent = True
-            continue
-
-        # Check which node fields have changed
-        changed_fields = [field for field in ['status', 'bond', 'slash_points'] if
-                          local_node[field] != remote_node[field]]
-
-        # Check if there are any changes
-        if len(changed_fields) > 0:
-            text = 'THORNode: ' + address + '\n' + \
-                   'Status: ' + local_node['status'].capitalize() + \
-                   ' ➡️ ' + remote_node['status'].capitalize() + '\n' + \
-                   'Bond: ' + tor_to_rune(int(local_node['bond'])) + \
-                   ' ➡️ ' + tor_to_rune(int(remote_node['bond'])) + '\n' + \
-                   'Slash Points: ' + '{:,}'.format(int(local_node['slash_points'])) + \
-                   ' ➡️ ' + '{:,}'.format(int(remote_node['slash_points']))
-
-            # Update data
-            local_node['status'] = remote_node['status']
-            local_node['bond'] = remote_node['bond']
-            local_node['slash_points'] = remote_node['slash_points']
-
-            # Send message
-            context.bot.send_message(chat_id, text)
-            message_sent = True
-
-    for address in delete_addresses:
-        del user_data['nodes'][address]
-
-    if message_sent:
-        show_home_menu(context=context, chat_id=chat_id)
-
-
-def check_thorchain_block_height(context):
-    """
-    Make sure the block height increases
-    """
-
-    chat_id = context.job.context['chat_id']
-    user_data = context.job.context['user_data']
-
-    block_height = get_thorchain_block_height()
-
-    # Check if block height got stuck
-    if 'block_height' in user_data and block_height <= user_data['block_height']:
-
-        # Increase stuck count to know if we already sent a notification
-        user_data['block_height_stuck_count'] += 1
-    else:
-        # Check if we have to send a notification that the Height increases again
-        if 'block_height_stuck_count' in user_data and user_data['block_height_stuck_count'] > 0:
-            text = 'Block height is increasing again! 👌' + '\n' + \
-                   'IP: ' + THORCHAIN_NODE_IP + '\n' + \
-                   'Block height now at: ' + block_height + '\n'
-            context.bot.send_message(chat_id, text)
-            user_data['block_height_stuck_count'] = -1
-        else:
-            user_data['block_height_stuck_count'] = 0
-
-    # Set current block height
-    user_data['block_height'] = block_height
-
-    # If it just got stuck send a message
-    if user_data['block_height_stuck_count'] == 1:
-        text = 'Block height is not increasing anymore! 💀' + '\n' + \
-               'IP: ' + THORCHAIN_NODE_IP + '\n' + \
-               'Block height stuck at: ' + block_height + '\n\n' + \
-               'Please check your Thornode immediately!'
-        context.bot.send_message(chat_id, text)
-
-    # Show buttons if there were changes or block height just got (un)stuck
-    # Stuck count:
-    # 0 == everthings alright
-    # 1 == just got stuck
-    # -1 == just got unstuck
-    # > 1 == still stuck
-
-    if user_data['block_height_stuck_count'] == 1 or user_data['block_height_stuck_count'] == -1:
-        show_home_menu(context=context, chat_id=chat_id)
-
-
-def check_thorchain_catch_up_status(context):
-    """
-    Check if node is some blocks behind with catch up status
-    """
-
-    chat_id = context.job.context['chat_id']
-    user_data = context.job.context['user_data']
-
-    if 'is_catching_up' not in user_data:
-        user_data['is_catching_up'] = False
-
-    is_currently_catching_up = is_thorchain_catching_up()
-    if user_data['is_catching_up'] == False and is_currently_catching_up:
-        user_data['is_catching_up'] = True
-        text = 'The Node is behind the latest block height and catching up! 💀 ' + '\n' + \
-               'IP: ' + THORCHAIN_NODE_IP + '\n' + \
-               'Current block height: ' + get_thorchain_block_height() + '\n\n' + \
-               'Please check your Thornode immediately!'
-        context.bot.send_message(chat_id, text)
-        show_home_menu(context=context, chat_id=chat_id)
-    elif user_data['is_catching_up'] == True and not is_currently_catching_up:
-        user_data['is_catching_up'] = False
-        text = 'The node caught up to the latest block height again! 👌' + '\n' + \
-               'IP: ' + THORCHAIN_NODE_IP + '\n' + \
-               'Current block height: ' + get_thorchain_block_height()
-        context.bot.send_message(chat_id, text)
-        show_home_menu(context=context, chat_id=chat_id)
-
-
-def check_thorchain_midgard_api(context):
-    """
-    Check that Midgard API is ok
-    """
-
-    chat_id = context.job.context['chat_id']
-    user_data = context.job.context['user_data']
-
-    if 'is_midgard_healthy' not in user_data:
-        user_data['is_midgard_healthy'] = True
-
-    is_midgard_currently_healthy = is_thorchain_midgard_healthy()
-    if user_data['is_midgard_healthy'] == True and not is_midgard_currently_healthy:
-        user_data['is_midgard_healthy'] = False
-        text = 'Midgard API is not healthy anymore! 💀' + '\n' + \
-               'IP: ' + THORCHAIN_NODE_IP + '\n\n' + \
-               'Please check your Thornode immediately!'
-        context.bot.send_message(chat_id, text)
-        show_home_menu(context, chat_id=chat_id)
-    elif user_data['is_midgard_healthy'] == False and is_midgard_currently_healthy:
-        user_data['is_midgard_healthy'] = True
-        text = 'Midgard API is healthy again! 👌' + '\n' + \
-               'IP: ' + THORCHAIN_NODE_IP + '\n'
-        context.bot.send_message(chat_id, text)
-        show_home_menu(context, chat_id=chat_id)
-
-
-def update_health_check_file(context):
-    """
-    Write timestamp into health.check file for the health check
-    """
-
-    with open('storage/health.check', 'w') as healthcheck_file:
-        timestamp = datetime.timestamp(datetime.now())
-        healthcheck_file.write(str(timestamp))
-
-
-"""
-######################################################################################################################################################
-Helpers
-######################################################################################################################################################
-"""
-
-
-def show_home_menu(context, chat_id, query=None):
-    """
-    Show buttons of home menu
-    """
-
-    keyboard = [[InlineKeyboardButton('My THORNodes', callback_data='thornode_menu'),
-                 InlineKeyboardButton('Admin Area', callback_data='admin_menu')]]
-
-    text = 'I am your THORNode Bot. 🤖\nChoose an action:'
-    # Edit message or write a new one depending on function call
-    if query:
-        query.edit_message_text(text,
-                                reply_markup=InlineKeyboardMarkup(keyboard))
-    else:
-        context.bot.send_message(chat_id, text,
-                                 reply_markup=InlineKeyboardMarkup(keyboard))
-
-
-def show_thornode_menu(context, chat_id, user_data, query=None):
-    """
-    Show buttons for supported actions.
-    """
-
-    keyboard = [[]]
-
-    for address in user_data['nodes'].keys():
-        keyboard.append([InlineKeyboardButton(address, callback_data='thornode_details-' + address)])
-
-    keyboard.append([InlineKeyboardButton('Add THORNode', callback_data='add_thornode'),
-                     InlineKeyboardButton('<< Back', callback_data='back_button')])
-
-    # Edit query message. Write a new message instead after address input
-    if query:
-        query.edit_message_text('Choose an address from the list below or add one:',
-                                reply_markup=InlineKeyboardMarkup(keyboard))
-    else:
-        context.bot.send_message(chat_id, 'Choose an address from the list below or add one:',
-                                 reply_markup=InlineKeyboardMarkup(keyboard))
-
-
-def show_detail_menu(query, address):
-    """
-    Show detail buttons for selected address
-    """
-
-    keyboard = [[
-        InlineKeyboardButton('Show THORNode Stats', callback_data='show_stats'),
-        InlineKeyboardButton('Delete THORNode', callback_data='confirm_thornode_deletion')
-        ],
-        [
-            InlineKeyboardButton('<< Back', callback_data='back_button')
-        ]]
-
-    # Send message
-    text = "You chose\n" + address + "\nWhat do you want to do with that Node?"
-    query.edit_message_text(text, reply_markup=InlineKeyboardMarkup(keyboard))
-    return WAIT_FOR_DETAIL
-
-
-def show_admin_menu(context, chat_id, query=None):
-    """
-    Show buttons of admin area
-    """
-
-    containers = get_running_docker_container()
-    if containers == "ERROR":
-        if query:
-            query.answer("Error while getting running docker container", show_alert=True)
-        print("Error while getting running docker container")
-        return END
-
-    # build keyboard with one button for every container
-    keyboard = [[]]
-    for container in containers:
-        for name in container['Names']:
-            container_name = name.replace('/', '')
-            status = container['Status']
-            text = container_name + " - " + status
-            keyboard.append([InlineKeyboardButton(text, callback_data='container-#' + container_name)])
-
-    keyboard.append([InlineKeyboardButton('<< Back', callback_data='back_button')])
-
-    # Send message
-    text = "⚠️ You're in the Admin Area - proceed with care ⚠️\n" \
-           "Below is a list of docker containers running on your system.\n" \
-           "Click on any container to restart it!"
-    if query:
-        query.answer()
-        query.edit_message_text(text, reply_markup=InlineKeyboardMarkup(keyboard))
-    else:
-        context.bot.send_message(chat_id, text, reply_markup=InlineKeyboardMarkup(keyboard))
-
-
-def get_running_docker_container():
-    """
-    Return Json of all running container on the host machine
-    """
-
-    bash_command = DOCKER_CURL_CMD + " http://localhost/containers/json"
-    process = subprocess.Popen(bash_command.split(), stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    output, error = process.communicate()
-    rc = process.returncode
-
-    if error or rc:
-        print(error)
-        return "ERROR"
-
-    container_string = output.decode('utf8')
-    return json.loads(container_string)
-
-
-def get_thornode_object(address):
-    """
-    Query nodeaccounts endpoints and return the Thornode object
-    """
-
-    while True:
-        response = requests.get(url=get_thorchain_validators())
-        if response.status_code == 200:
-            break
-
-    nodes = response.json()
-
-    # Get the right node
-    node = next(filter(lambda node: node['node_address'] == address, nodes), None)
-    return node
-
-
-def get_thorchain_block_height():
-    """
-    Return block height of your Thornode
-    """
-
-    while True:
-        response = requests.get(url=get_thorchain_status())
-        if response.status_code == 200:
-            break
-
-    status = response.json()
-    return status['result']['sync_info']['latest_block_height']
-
-
-def is_thorchain_catching_up():
-    response = requests.get(url=get_thorchain_status())
-    if response.status_code != 200:
-        return True
-
-    status = response.json()
-    return status['result']['sync_info']['catching_up']
-
-
-def is_thorchain_midgard_healthy():
-    """
-    Return status of Midgard API
-    """
-
-    response = requests.get(url=get_thorchain_midgard_endpoint())
-    if response.status_code != 200:
-        return False
-
-    if response.text.find('"OK"') != -1:
-        return True
-    else:
-        return False
-
-
-def get_thorchain_validators():
-    """
-    Return the nodeaccounts endpoint to query data from.
-    """
-
-    return 'http://localhost:8000/nodeaccounts.json' if DEBUG else 'http://' + THORCHAIN_NODE_IP + ':1317/thorchain/nodeaccounts'
-
-
-def get_thorchain_status():
-    """
-    Return the endpoint for block height checks
-    """
-
-    return 'http://localhost:8000/status.json' if DEBUG else 'http://' + THORCHAIN_NODE_IP + ':26657/status'
-
-
-def get_thorchain_midgard_endpoint():
-    """
-    Return the endpoint for Midgard API check
-    """
-
-    return 'http://localhost:8000/midgard.json' if DEBUG else 'http://' + THORCHAIN_NODE_IP + ':8080/v1/health'
-
-
-def tor_to_rune(tor):
-    """
-    1e8 Tor are 1 Rune
-    Format depending if RUNE > or < Zero
-    """
-
-    if tor >= 100000000:
-        return "{:,} RUNE".format(int(tor / 100000000))
-    else:
-        return '{:.8f} RUNE'.format(tor / 100000000)
-
-
-def error(update, context):
-    """
-    Log error.
-    """
-
-    logger.warning('Update "%s" caused error: %s', update, context.error)
-
-
-"""
-######################################################################################################################################################
 Application
 ######################################################################################################################################################
 """
-
-# Shortcut for ConversationHandler.END
-END = ConversationHandler.END
-
-# Conversation state(s)
-THORNODE_MENU, WAIT_FOR_ADDRESS, WAIT_FOR_DETAIL, WAIT_FOR_CONFIRMATION, ADMIN_MENU = range(5)
 
 
 def main():
@@ -792,7 +356,7 @@ def main():
     bot = Updater(TELEGRAM_BOT_TOKEN, persistence=PicklePersistence(filename='storage/session.data'), use_context=True)
     dispatcher = bot.dispatcher
 
-    # Restart jobs for all users
+    # Restart jobs for all users helpers
     chat_ids = dispatcher.user_data.keys()
     for chat_id in chat_ids:
         dispatcher.job_queue.run_repeating(thornode_checks, interval=30, context={
